Fix torch_xla import error
Added an environment variable (USE_TORCH_XLA = "0") to disable XLA (Accelerated Linear Algebra) support in transformers. This prevents the import error with torch_xla where a symbol is undefined, which was likely caused by a version mismatch between PyTorch and torch_xla.

Since the script is intended to run on GPUs and not TPUs, disabling XLA is a safe solution.