Fix TypeError in LoraConfig initialization
Removed the invalid 'label_names' parameter from LoraConfig initialization in the Train-Phi3-4bit.py script. The LoraConfig class from the PEFT library doesn't accept this parameter, which was causing the TypeError. Removing this parameter resolves the issue without affecting other functionality, as it wasn't being used elsewhere in the code.